Move authentication module to token
