+++ lib/openid_token_proxy.rb
@@ -4,11 +4,11 @@ require 'openid_connect'
 
 require 'openid_token_proxy/error'
 
-require 'openid_token_proxy/authentication'
 require 'openid_token_proxy/client'
 require 'openid_token_proxy/config'
 require 'openid_token_proxy/engine'
 require 'openid_token_proxy/token'
+require 'openid_token_proxy/token/authentication'
 require 'openid_token_proxy/version'
 
 module OpenIDTokenProxy
